Extract generated test service code from django example.
